Coloquei tratemento de excecoes

+++ src/main/java/com/example/CrudFlix/Service/FilmeService.java
@@ -1,11 +1,14 @@
 package com.example.CrudFlix.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+
 import com.example.CrudFlix.Repository.FilmeRepository;
 import com.example.CrudFlix.Model.Filme;
+import com.example.CrudFlix.Exception.FilmeNaoEncontradoException;
 
 @Service
 public class FilmeService {
@@ -16,19 +19,37 @@ public class FilmeService {
         this.filmeRepository = filmeRepository;
     }
 
+
+    //Crud
+
     public List<Filme> listarFilmes(){
         return filmeRepository.findAll();
+    }
+
+    public Filme buscarFilmePorId(Long id) {
+        return filmeRepository.findById(id)
+            .orElseThrow(() -> new FilmeNaoEncontradoException(id));
     }
 
     public Filme salvarFilme(Filme filme){
         return filmeRepository.save(filme);
     }
 
-    public Filme buscarFilmePorId(Long id) {
-        return filmeRepository.findById(id).orElse(null);
+   public void atualizarFilme(Long id, Filme filmeAtualizado){
+        Filme filmeExistente = buscarFilmePorId(id);
+        if(filmeExistente == null) {
+            throw new FilmeNaoEncontradoException(id);
+        }
+        filmeExistente.setTitulo(filmeAtualizado.getTitulo());
+        filmeExistente.setGenero(filmeAtualizado.getGenero());
+        filmeExistente.setAnoLancamento(filmeAtualizado.getAnoLancamento());
+        filmeRepository.save(filmeExistente);
     }
 
     public void deletarFilme(Long id){
+        if (!filmeRepository.existsById(id)) {
+            throw new FilmeNaoEncontradoException(id);
+        }
         filmeRepository.deleteById(id);
     }
 }
